Sorting things into the right folders.
