+++ src/getputadventures/ItemLibrary.java
@@ -1,4 +1,4 @@
-package getput.adventures;
+package getputadventures;
 
 import java.io.File;
 import java.util.Scanner;
